transport/libfabric: Optimize Progress
Attempt to request FI_PROGRESS_AUTO to see if the libfabric provider
supports it, if it doesn't fall back to FI_PROGRESS_MANUAL.
FI_PROGRESS_AUTO means that we do not need to call into the progress
engine for submitted operations to complete.  This means that we can
remove the host endpoint from the progress call, and we only need to
progress the host endpoint when user calls nvshmem_quiet() from the
host. This allows us to set the threading model as FI_THREAD_COMPELTION
because the host only progress the host EP, and the proxy only
progresses the proxy EP, leading to compliance with FI_THREAD_COMPLETION.
An edge case exists here where the user calls nvshmem_quiet() on the
host QP_IDX from a GPU kernel, but this is illegial because the user
shouldn't be calling QP API's on QP's not provided to them via the qp
creation API's.

This patch should offer a performance improvement because it reduces the
number of EP's that are progressed in the critical path, and it allows
the libfabric provider to reduce locking b/c of threading model
FI_THREAD_COMPLETION.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>